feat: Добавит аутентификацию пользователей и разграничение прав в соответствии с ролями.

+++ src/controllers/user-controller.ts
@@ -1,11 +1,19 @@
 import type { Request, Response } from 'express';
 import prisma from '../prisma/prisma';
-import type { CreateUser, UpdateUser } from '../types/user-interface';
+import { UserRole, type CreateUser, type UpdateUser } from '../types/user-interface';
 
 export const userController = {
   getUsers: async (req: Request, res: Response) => {
     try {
-      const users = await prisma.user.findMany();
+      const users = await prisma.user.findMany({
+        select: {
+          id: true,
+          name: true,
+          email: true,
+          role: true,
+          status: true,
+        },
+      });
       res.json(users);
     } catch (error) {
       console.error(error);
@@ -18,30 +26,27 @@ export const userController = {
       const { id } = req.params;
       const user = await prisma.user.findUnique({
         where: { id },
+        select: {
+          id: true,
+          name: true,
+          email: true,
+          role: true,
+          status: true,
+        },
       });
 
       if (!user) {
         return res.status(404).json({ error: `Пользователь с id: ${id} не найден` });
       }
+
+      if (req.user?.role !== UserRole.ADMIN && req.user?.userId !== id) {
+        return res.status(403).json({ error: 'Недостаточно прав.' });
+      }
+
       res.json(user);
     } catch (error) {
       console.error(error);
       res.status(500).json({ error: 'Ошибка сервера. Мы уже бежим исправлять это!' });
-    }
-  },
-
-  createUser: async (req: Request, res: Response) => {
-    try {
-      const data = req.body as CreateUser;
-
-      const user = await prisma.user.create({
-        data,
-      });
-
-      res.status(201).json(user);
-    } catch (error) {
-      console.error(error);
-      res.status(400).json({ error: 'Ошибка при создании пользователя.' });
     }
   },
 
@@ -50,12 +55,24 @@ export const userController = {
       const { id } = req.params;
       const data = req.body as UpdateUser;
 
+      console.log(req.body);
+
+      if (req.user?.role !== UserRole.ADMIN && req.user?.userId !== id) {
+        return res.status(403).json({ error: 'Недостаточно прав.' });
+      }
+
       const user = await prisma.user.update({
         where: { id },
         data,
       });
 
-      res.json(user);
+      res.json({
+        id: user.id,
+        name: user.name,
+        email: user.email,
+        role: user.role,
+        status: user.status,
+      });
     } catch (error) {
       console.error(error);
       res.status(400).json({ error: 'Ошибка при обновлении пользователя.' });
@@ -65,6 +82,14 @@ export const userController = {
   deleteUser: async (req: Request, res: Response) => {
     try {
       const { id } = req.params;
+
+      if (req.user?.role !== UserRole.ADMIN && req.user?.role !== id) {
+        return res.status(403).json({ error: 'Недостаточно прав.' });
+      }
+
+      if (req.user?.userId === id) {
+        return res.status(400).json({ error: 'Нельзя удалить свой аккаунт.' });
+      }
 
       await prisma.user.delete({
         where: { id },
